+++ embedding/providers_test.go
@@ -83,7 +83,7 @@ func TestOllamaEmbedding(t *testing.T) {
 
 		embedding, err := e.GetTextEmbedding(context.Background(), "test text")
 		require.NoError(t, err)
-		assert.Equal(t, []float64{0.1, 0.2, 0.3, 0.4, 0.5}, embedding)
+		assert.Equal(t, []float32{0.1, 0.2, 0.3, 0.4, 0.5}, embedding)
 	})
 
 	t.Run("GetQueryEmbedding with mock server", func(t *testing.T) {
@@ -197,7 +197,7 @@ func TestCohereEmbedding(t *testing.T) {
 
 		embedding, err := e.GetTextEmbedding(context.Background(), "test text")
 		require.NoError(t, err)
-		assert.Equal(t, []float64{0.1, 0.2, 0.3}, embedding)
+		assert.Equal(t, []float32{0.1, 0.2, 0.3}, embedding)
 	})
 
 	t.Run("GetQueryEmbedding uses search_query input type", func(t *testing.T) {
@@ -334,7 +334,7 @@ func TestHuggingFaceEmbedding(t *testing.T) {
 
 		embedding, err := e.GetTextEmbedding(context.Background(), "test text")
 		require.NoError(t, err)
-		assert.Equal(t, []float64{0.1, 0.2, 0.3, 0.4}, embedding)
+		assert.Equal(t, []float32{0.1, 0.2, 0.3, 0.4}, embedding)
 	})
 
 	t.Run("GetTextEmbedding with TEI mock", func(t *testing.T) {
@@ -430,13 +430,13 @@ func TestMeanPool(t *testing.T) {
 			{3.0, 6.0, 9.0},
 		}
 		result := meanPool(tokens)
-		assert.Equal(t, []float64{2.0, 4.0, 6.0}, result)
+		assert.Equal(t, []float32{2.0, 4.0, 6.0}, result)
 	})
 
 	t.Run("Handles single token", func(t *testing.T) {
 		tokens := [][]float64{{1.0, 2.0, 3.0}}
 		result := meanPool(tokens)
-		assert.Equal(t, []float64{1.0, 2.0, 3.0}, result)
+		assert.Equal(t, []float32{1.0, 2.0, 3.0}, result)
 	})
 
 	t.Run("Handles empty input", func(t *testing.T) {
