+++ index/summary.go
@@ -396,16 +396,16 @@ func (r *SummaryIndexRetriever) retrieveLLM(ctx context.Context, query schema.Qu
 }
 
 // cosineSimilarity calculates the cosine similarity between two vectors.
-func cosineSimilarity(a, b []float64) float64 {
+func cosineSimilarity(a, b []float32) float64 {
 	if len(a) != len(b) {
 		return 0
 	}
 
 	var dotProduct, normA, normB float64
 	for i := range a {
-		dotProduct += a[i] * b[i]
-		normA += a[i] * a[i]
-		normB += b[i] * b[i]
+		dotProduct += float64(a[i]) * float64(b[i])
+		normA += float64(a[i]) * float64(a[i])
+		normB += float64(b[i]) * float64(b[i])
 	}
 
 	if normA == 0 || normB == 0 {
